skipping html generation if no version was updated
